Add a "monitor" button
When enabled, the application will check the input media directory for new files at every interval and if any are added, they will be appended to the bottom of the media list. Any deletions will be removed. Newly appended media will immediately selected as the active media.

The intention for this feature is for setting the output folder for a Stable Diffusion as the target media folder, or something similar. As new images or videos are generated, Rope will pick those items up and display them using the swap settings you last used. Just keep it off if you want Rope to function as normal.